Move rule_set.rs to rules/mod.rs
