+++ src/rules/code_style/mod.rs
@@ -1,3 +1,2 @@
-pub mod whitespace;
 pub mod line_length;
-
+pub mod whitespace;
